py/requirements.txt
asyncpg==0.12.0
av==0.3.3
cycler==0.10.0
decorator==4.1.2
matplotlib==2.0.2
networkx==1.11
numpy==1.13.1
olefile==0.44
Pillow==4.2.1
PIMS==0.4
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyWavelets==0.5.2
scikit-image==0.13.0
scipy==0.19.1
six==1.10.0
slicerator==0.9.8
